Ensure that MetadataDTO doesn't accidentally lazy load plugin info
We have the DTOs as a representation of a somewhat stripped down entity that
comes without its references. This avoids loading a load of more data if, for
example, it gets serialized, which is particularly important since we disabled
JPA in views. The plugin reference in the MetadataDTO could potentially still
cause such a lazy loading issue and there was no real reason it was necessary.
As such, it was changed to only contain the ID, which should not cause any
unnecessary loads.