Prevent creation & update of a smart playlist if the query is invalid
A smart playlist is based upon a query that will run every time to find the
videos that belong to it. Having an error in the query would thus always cause
an error when loading the videos associated with the playlist. It will now be
checked before the smart playlist is created or updated.