Make sure we synchronize the version to the frontend
So far the frontend version was disconnected from the backend version. This can
(and has) caused the versions to be different from another because it might be
forgotten to update in both places. To make sure this does not happen, this now
sets up a task the synchronizes the version in the frontend before packaging.