Make tolerance for duration queries configurable
When doing a "like" query for duration fields, the tolerance was so far static
and set to 3 seconds or minutes depending on the specified duration. If it was
longer than an hour, the tolerance was 3 minutes (e.g. `Length~1h15m`). Any
shorter durations had a tolerance of 3 seconds. This `3` value is now adjustable
 in the application properties with `duration-query-tolerance` and it just
defaults to 3. It still has the behavior of applying either to minutes or
seconds depending on the duration itself.