Base tolerance on percentage instead
The tolerance was previously configured to be in minutes for hour durations and
seconds for minute durations. This has two problems: for values that are close
to border between the two it could either be too much or too little.
Additionally, this does not take into account longer durations, though those are
 not supported by the input either.

This now switches the tolerance to be a percentage of the specified value
instead, which should scale to any length.

The default has thus been adjusted to `5` meaning `5%` which is roughly what it
was before (though obviously not the exact same).