Replace vue-cli with vite
Starting up the vue cli dev server (i.e. webpack) often took quite some time,
somewhere around a minute. Vue-cli also resulted in a very deep dependency tree
that resulted in a lot of possible vulnerabilities without a way to update them
on the user side. This wasn't that much of a concern given that these were
issues with the development tools and not the code that will be included in the
production build, but one thing could always lead to the other.